Added persistance to assessments and averages

+++ app/src/main/java/com/nodomain/mark/class_tool/CourseListAdapter.java
@@ -44,7 +44,7 @@ public class CourseListAdapter extends ArrayAdapter<course>{
                 averageText = "";
             }
             else {
-                averageText = currentCourse.average + "%";
+                averageText = (int)currentCourse.average + "%";
             }
             gradeDisp.setText(averageText);
         }
